bd change

+++ database/migrations/2024_06_30_033438_create_inventories_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('inventories', function (Blueprint $table) {
             $table->id();
             $table->string('name', 45);
-            $table->integer('amount');
+            $table->double('amount', 10, 2);
             $table->enum('type', ['Verdura', 'Fruta', 'Proteina', 'Cereales y Legumbres'])->default('Verdura');
             $table->timestamps();
         });
